requirements.txt
python-twitter==3.5
requests==2.25.1
